requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.16.0
dash-bootstrap-components==1.5.0
dash-bootstrap-templates==1.1.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.2
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.1
plotly==5.19.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
six==1.16.0
tenacity==8.2.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
